unify XObjectEventConverter
